Allow root App component to have its own preFetch
Sometimes the App requires global preFetch. An example of that could be the case when you have an external admin panel, providing edit interface for some amount of variables, and a single api method, returning a whole snapshot of this data. In this case the whole frontend app requires just a single prefetch, which is not to be placed in any particular route-dependent component, but rather to be done somehow globally.